aula2-datalake-aws/emr/requirements.txt
boto3==1.21.0
python-dotenv==0.19.2